requirements.txt
################################################################################
#                                                                              #
#                           Requirements for testing                           #
#    (If you need production dependencies, add them to the `setup.py` file)    #
#                                                                              #
################################################################################

django
django-cas-ng
django-cors-headers
django-debug-toolbar
django-filter
django-minio-storage
django-prometheus
django-silk
django-uwsgi
djangorestframework
psycopg2-binary
pygments
python-json-logger
requests


# test libraries
coverage
flake8
flake8-quotes
tblib
